Moved Tests to an API folder to avoid errors when trying to find the database
Removed Order of tests from User Tests
Added 2 tests that suceed if they fail, one for blank username/password and another for one over 20 characters
