+++ src/test/java/dev/group4/api/PotLuckServiceTests.java
@@ -1,4 +1,4 @@
-package dev.group4;
+package dev.group4.api;
 
 import dev.group4.aspects.InvalidTimeException;
 import dev.group4.entities.Potluck;
